review comments: clean up
